+ арзделение js-кода на модули
