+++ js/main.js
@@ -1,89 +1,2 @@
-const DESCRIPTIONS = ['Утро', 'Котик', 'Солнышко', 'Дерево'];
-const LIKES_MIN = 15;
-const LIKES_MAX = 200;
-const COMMENTS_MIN = 0;
-const COMMENTS_MAX = 30;
-const AVATAR_MIN = 1;
-const AVATAR_MAX = 6;
-const PHOTOS_MIN = 1;
-const PHOTOS_MAX = 25;
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-const NAMES = ['Иван', 'Пётр', 'Александр', 'Наталья', 'Екатерина', 'Татьяна', 'Олег', 'Владимир', 'Михаил', 'Константин'];
-
-const getPositiveRandomInteger = (a, b) => {
-  const lowerLimit = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
-  const upperLimit = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
-  const positiveRandomInteger = Math.random() * (upperLimit - lowerLimit + 1) + lowerLimit;
-  return Math.floor(positiveRandomInteger);
-};
-
-const getUnique = (a, b) => {
-  const previousValues = [];
-
-  return function () {
-    let currentValue = getPositiveRandomInteger(a, b);
-
-    if (previousValues.length >= (b - a + 1)) {
-      console.error('Перебраны все числа из диапазона от ' + a + ' до ' + b);
-      return null;
-    }
-
-    while (previousValues.includes(currentValue)) {
-      currentValue = getPositiveRandomInteger(a, b);
-    }
-    previousValues.push(currentValue);
-    return currentValue;
-  };
-};
-
-const getUniqueId = getUnique(PHOTOS_MIN, PHOTOS_MAX);
-const getUniquPhoto = getUnique(PHOTOS_MIN, PHOTOS_MAX);
-
-const getRandomElement = (arr) => arr[getPositiveRandomInteger(0, arr.length - 1)];
-
-const getComment = () => ({
-  id: getPositiveRandomInteger(COMMENTS_MIN, COMMENTS_MAX),
-  avatar: `img/avatar-${getPositiveRandomInteger(AVATAR_MIN, AVATAR_MAX)}.svg`,
-  message: getRandomElement(MESSAGES),
-  name: getRandomElement(NAMES),
-});
-
-
-const getComments = () => {
-  const count = getPositiveRandomInteger(COMMENTS_MIN, COMMENTS_MAX);
-  const comments = [];
-  for (let i = 1; i <= count; i++) {
-    comments.push(getComment());
-  }
-  return comments;
-};
-
-const getPhoto = () => ({
-  id: getUniqueId(),
-  url: `photos/${getUniquPhoto()}.jpg`,
-  description: getRandomElement(DESCRIPTIONS),
-  likes: getPositiveRandomInteger(LIKES_MIN, LIKES_MAX),
-  comments: getComments()
-});
-
-
-// const createPhotos = () => {
-//   const result = [];
-
-//   for (let i = 1; i <= 25; i++) {
-//     result.push(getPhoto())
-//   }
-
-//   return result;
-// };
-
-const createPhotos = () => Array.from({ length: PHOTOS_MAX }, getPhoto);
-
-console.log(createPhotos());
+import './util.js';
+import './data.js';
